requirements.txt
ultralytics
opencv-python
numpy
pillow
pyyaml
torch
torchvision
